# ms-gateway/docs/swagger.yaml
definitions:
  helper.Message:
    properties:
      message:
        type: string
    type: object
  helper.MessageDetails:
    properties:
      detail:
        type: string
      message:
        type: string
    type: object
  model.Address:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: integer
    type: object
  model.AddressRequest:
    properties:
      city:
        type: string
      country:
        type: string
    type: object
...
  model.InfoCustomerResponse:
    properties:
      address:
        $ref: '#/definitions/model.AddressRequest'
      user:
        $ref: '#/definitions/model.RegisterResponse'
    type: object
  model.InputRegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
...
  model.RegisterResponse:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: arviansyahnur3@gmail.com
    name: Arviansyah Nur
    url: http://www.swagger.io/support
  description: An e-commerce API for booking hotel and with many features, made using
    microservices and gRPC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: vihotel
  version: "1.0"
paths:
...
  /api/v1/user:
    delete:
      consumes:
      - application/json
      description: Delete customer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Customer
    get:
      consumes:
      - application/json
      description: Get info customer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.InfoCustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Update info customer
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.InputRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Customer
  /api/v1/user/address:
    post:
      consumes:
      - application/json
      description: Add address customer
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Address'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Update address customer
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Address'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Customer
  /api/v1/user/admin:
    get:
      consumes:
      - application/json
      description: Get all customer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Admin
  /api/v1/user/admin/:id:
    delete:
      consumes:
      - application/json
      description: Delete customer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helper.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get info customer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update info customer
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.InputRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Admin
  /login:
    post:
      consumes:
      - application/json
      description: Login email and password
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Public
      tags:
      - User
  /register:
    post:
      consumes:
      - application/json
      description: Register name, email, and password
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.InputRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Public
      tags:
      - User
swagger: "2.0"